refactor: Coupon 패키지 분리
