+++ src/main/java/coupon/domain/coupon/Category.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.domain.coupon;
 
 public enum Category {
     FASHION,
