+++ src/main/java/coupon/domain/coupon/DiscountAmount.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.domain.coupon;
 
 import jakarta.persistence.Embeddable;
