+++ src/main/java/coupon/service/CouponService.java
@@ -5,7 +5,7 @@ import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import coupon.domain.Coupon;
+import coupon.domain.coupon.Coupon;
 import coupon.repository.CouponRepository;
 
 @Service
