+++ src/test/java/coupon/domain/coupon/DiscountRateTest.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.domain.coupon;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -8,6 +8,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import coupon.domain.coupon.DiscountAmount;
+import coupon.domain.coupon.DiscountRate;
+import coupon.domain.coupon.MinOrderPrice;
 
 class DiscountRateTest {
