+++ src/test/java/coupon/domain/coupon/NameTest.java
@@ -1,14 +1,15 @@
-package coupon.domain;
+package coupon.domain.coupon;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.apache.logging.log4j.util.Strings;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import coupon.domain.coupon.Name;
 
 class NameTest {
