+++ src/test/java/coupon/service/CouponServiceTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import coupon.domain.Category;
-import coupon.domain.Coupon;
+import coupon.domain.coupon.Category;
+import coupon.domain.coupon.Coupon;
 
 @SpringBootTest
 class CouponServiceTest {
